refactor: extract WslGatewayKeepAliveService from App.xaml.cs
Move the WSL gateway keepalive logic (the fire-and-forget task that keeps
the local gateway's WSL2 VM alive after the tray exits, plus stale-process
and stale-marker cleanup) out of App.xaml.cs into a dedicated service.

App.xaml.cs kept nine private methods (~220 lines) for this concern. They
are moved verbatim into Services/WslGatewayKeepAliveService.cs; the only
adaptation is that the two App fields they read, _settings and
_gatewayRegistry, now enter as Func<> delegates instead of direct field
access. No logic, ordering, threading, or side effect changes.

The keep-alive *window* (_keepAliveWindow, the WinUI XamlRoot anchor) is a
separate concern and stays in App.

Seam unchanged: OnLaunchedAsync still fires the work via Task.Run before
InitializeGatewayClient(), off the UI thread.
